+++ qiskit/__init__.py
@@ -23,8 +23,8 @@ from .exceptions import QiskitError
 from qiskit.circuit import ClassicalRegister
 from qiskit.circuit import QuantumRegister
 from qiskit.circuit import QuantumCircuit
-from qiskit.tools.compiler import (compile, execute)  # TODO remove after 0.8
-from qiskit.execute_wrapper import execute_circuits
+from qiskit.tools.compiler import compile  # TODO remove after 0.8
+from qiskit.execute import (execute_circuits, execute)
 
 # The qiskit.extensions.x imports needs to be placed here due to the
 # mechanism for adding gates dynamically.
